% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{setup-helpers}
\alias{setup-helpers}
\alias{use_coc}
\alias{use_contributing}
\alias{use_jdt_readme}
\alias{use_jdtdown}
\title{Helpers for project setup with jdtdown}
\usage{
use_coc()

use_contributing()

use_jdt_readme()

use_jdtdown(config_file = "_pkgdown.yml", destdir = "reference")
}
\arguments{
\item{config_file}{Path to the pkgdown yaml config file - could be set to be
\verb{pkgdown/} subfolder.}

\item{destdir}{Target directory for pkgdown docs. By default, it will be in
\code{reference} sub directory for R Markdown related package using \strong{jdtdown}.}
}
\description{
These helper are mostly inspired by the one in \strong{usethis} from the tidyverse
team and are useful to setup a new pkgdown project with \strong{jdtdown}
}
\details{
\itemize{
\item \code{use_coc()}: put _CODE\_OF\\emph{CONDUCT.md} in \verb{.github/} directory setting the
contact email to jdt's.
\item \code{use_contributing()}: Add a \emph{CONTRIBUTING.md} file in \verb{.github/} following a
template in \strong{jdtdown}. Inspired by \code{usethis::use_tidy_contributing()}.
\item \code{use_jdt_readme()}: Add a \emph{README.md} file in the structure used by JDT's packages.
}
}
